修复添加接口日志后小程序授权失败问题.

+++ src/main/java/com/bkit/fatdown/service/impl/FoodElementServiceImpl.java
@@ -124,7 +124,7 @@ public class FoodElementServiceImpl implements IFoodElementService {
         TbFoodElementRelationExample example = new TbFoodElementRelationExample();
         example.createCriteria()
                 .andFoodEqualTo(foodId);
-        return (int)relationMapper.countByExample(example);
+        return relationMapper.countByExample(example);
     }
 
     /**
@@ -136,7 +136,7 @@ public class FoodElementServiceImpl implements IFoodElementService {
         TbFoodElementRelationExample example = new TbFoodElementRelationExample();
         example.createCriteria()
                 .andIdEqualTo(relationId);
-        return (int)relationMapper.countByExample(example);
+        return relationMapper.countByExample(example);
     }
 
     /**
